Updates to Nexus core while getting the scaling experiment running.
(*) Added specifying the working directory of your executor. If you
specify the empty string than it goes in the old default spot.

(*) When an executor dies or exits we no longer kill frameworks, now
we send them TASK_LOST messages. Note that this is still probably not
good enough, we probably want to add a M2F_LOST_EXECUTOR message and
the appropriate callback (executorLost).