Updated to protocol buffers. (This is a massive commit, for which I apologize.)
In addition to updating the code to use protocol buffers:

(1) A new Java implementation is included (because getting protocol
buffers to work was too hard so it was finally time to leave SWIG
behind).

(2) The C API has been dropped (and is waiting for someone who wants
to do it and maintain it).

(3) The Python API has been broken. We plan on supporting it, but not
right now.