Implemented FrameworksManager.
Refactored FramworksManager to separate interface and implementation.
Currently has implementations for the following:
--> public: list, add, remove,resurrect and exists
--> private: expire, cache

Implemented test cases for frameworks manager.

Tests all the methods of FrameworksManager class.
Mocked the FrameworksStorage class for tests.

Also, Added an initial FrameworksManager and a 'mesos' binary that can be used for communicating with a
running Mesos instance, for example, to submit a scheduler.